# openapi.yaml
openapi: 3.0.3
info:
  title: API сервиса для тестового задания
  version: 0.1.0
...
paths:
...
  /schools/{school_id}/classes/{class_id}/students:
    parameters:
      - $ref: '#/components/parameters/school_id'
      - $ref: '#/components/parameters/class_id'
    get:
      tags:
        - students
        - classes
      summary: Вывести список студентов класса
      operationId: getClassStudentList
      responses:
        '200':
          description: Список студентов
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Ожидается токен, выдаваемый при регистрации студента
      bearerFormat: SHA256
  parameters:
    school_id:
      name: school_id
      in: path
      required: true
      schema:
        type: integer
        format: int32
    class_id:
      name: class_id
      description: id класса
      in: path
      required: true
      schema:
        type: integer
        format: int32
  schemas:
...
    Student:
      required:
        - id
        - first_name
        - last_name
        - surname
        - class_id
        - school_id
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
          readOnly: true
        first_name:
          type: string
          example: Вячеслав
        last_name:
          type: string
          example: Абдурахмангаджиевич
        surname:
          type: string
          example: Мухобойников-Сыркин
        class_id:
          type: integer
          format: int32
        school_id:
          type: integer
          format: int32